+++ jm_serial_port_posix.cpp
@@ -537,7 +537,7 @@ int32 serial_port::set_rts(bool set) {
     return ioctl(_fd, TIOCMSET, &status);
 }
 
-int64 serial_port::read(uint8 *data, int64 offset, int64 count) {
+int64 serial_port::read(uint8 *data, int32 offset, int32 count) {
     if (data == NULL)
         return -1;
 
@@ -551,7 +551,7 @@ int64 serial_port::read(uint8 *data, int64 offset, int64 count) {
     return ret_val;
 }
 
-int64 serial_port::write(const uint8 *data, int64 offset, int64 count) {
+int64 serial_port::write(const uint8 *data, int32 offset, int32 count) {
     if (data == NULL)
         return -1;
